+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 * Fix #6627: Removed deprecated methods from `io.fabric8.kubernetes.client.CustomResource` class
 * Fix #6628: Removed deprecated `io.fabric8.kubernetes.client.CustomResourceList` class
 * Fix #6630: Removed deprecated `io.fabric8.kubernetes.client.utils.ReflectUtils` class
+* Fix #6455: (crd-generator) Changed format in `@PrinterColumn` to enum
 
 ### 6.13.4 (2024-09-25)